SUPR3HardenedMain: size_t
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@18499 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostDrivers/Support/SUPR3HardenedMain.cpp
@@ -393,7 +393,7 @@ DECLHIDDEN(int) supR3HardenedPathProgram(char *pszPath, size_t cchPath)
     /*
      * Calc the length and check if there is space before copying.
      */
-    unsigned cch = strlen(g_szSupLibHardenedDirPath) + 1;
+    size_t cch = strlen(g_szSupLibHardenedDirPath) + 1;
     if (cch <= cchPath)
     {
         memcpy(pszPath, g_szSupLibHardenedDirPath, cch + 1);
